Air Cargo Analysis Project 2
Course-End Project: Air Cargo Analysis


Problem Statement Scenario:
Air Cargo is an aviation company that provides air transportation services for
passengers and freight. Air Cargo uses its aircraft to provide different services with
the help of partnerships or alliances with other airlines. The company wants to
prepare reports on regular passengers, busiest routes, ticket sales details, and
other scenarios to improve the ease of travel and booking for customers.
 
Project Objective:
You, as a DBA expert, need to focus on identifying the regular customers to provide offers, analyze the busiest route which helps to increase the number of aircraft required and prepare an analysis to determine the ticket sales details. This will ensure that the company improves its operability and becomes more customer-centric and a favorable choice for air travel.